refactor for normalization
